Factor out test code for comparing fixtures
